moved about to bio, and then move bio and press under about menu
removed typesafe-actions; accidentally installed
