+++ app.js
@@ -46,6 +46,8 @@ app.get('/health-check', (req, res) => res.sendStatus(200));
 const oldRoutesToRedirectsMap = {
     '/home': '/',
     '/music': '/media/music',
+    '/about': '/about/bio',
+    '/press': '/about/press',
 };
 
 Object.keys(oldRoutesToRedirectsMap).forEach(key => (
