+++ server/src/meta.ts
@@ -20,7 +20,7 @@ const baseString = 'Sean Chen: Pianist, Composer, Arranger | ';
 const age = moment().diff('1988-08-27', 'year');
 const descriptions: {
     home: string;
-    about: string;
+    bio: string;
     contact: string;
     upcoming: string;
     archive: string;
@@ -33,7 +33,7 @@ const descriptions: {
     [index: string]: any;
 } = {
         home: 'Welcome to the official website of pianist, composer, and arranger Sean Chen. Third Prize at the 2013 Van Cliburn, Christel DeHaan Classical Fellow of the 2013 American Pianists Awards, and Artist-in-Residence at University of Missouri, Kansas City.',
-        about: `Hailed as a charismatic rising star with “an exceptional ability to connect with an audience combined with an easy virtuosity” (Huffington Post), ${age.toString()}-year-old American pianist Sean Chen, third prize winner at the 2013 Van Cliburn International Piano Competition and recipient of the DeHaan Classical Fellowship as the winner of the 2013 American Pianists Awards, has continued to earn accolades for “alluring, colorfully shaded renditions” (New York Times) and “genuinely sensitive” (LA Times) playing.`,
+        bio: `Hailed as a charismatic rising star with “an exceptional ability to connect with an audience combined with an easy virtuosity” (Huffington Post), ${age.toString()}-year-old American pianist Sean Chen, third prize winner at the 2013 Van Cliburn International Piano Competition and recipient of the DeHaan Classical Fellowship as the winner of the 2013 American Pianists Awards, has continued to earn accolades for “alluring, colorfully shaded renditions” (New York Times) and “genuinely sensitive” (LA Times) playing.`,
         contact: `Contact information for Sean Chen and for booking performances.`,
         upcoming: 'Upcoming recitals, concerti, and masterclasses.',
         archive: 'Past recitals, concerti, and masterclasses.',
@@ -46,7 +46,7 @@ const descriptions: {
         press: `Reviews of Sean Chen's performances.`,
     };
 
-const validFirst = [ '', 'about', 'contact', 'schedule', 'media', 'press' ];
+const validFirst = [ '', 'about', 'contact', 'schedule', 'media', 'press', 'bio', 'disc' ];
 
 export const getMetaFromPathAndSanitize = async (url: string) => {
     const parsed = regex.exec(url);
