+++ web/src/components/About/Bio/actionTypeKeys.ts
@@ -1,0 +1,7 @@
+enum BIO_ACTIONS {
+    FETCH_BIO_REQUEST = 'BIO--FETCH_BIO_REQUEST',
+    FETCH_BIO_SUCCESS = 'BIO--FETCH_BIO_SUCCESS',
+    FETCH_BIO_ERROR = 'BIO--FETCH_BIO_ERROR',
+}
+
+export default BIO_ACTIONS;
