+++ web/src/components/About/Bio/actionTypes.ts
@@ -1,0 +1,19 @@
+import BIO_ACTIONS from 'src/components/About/Bio/actionTypeKeys';
+import { Blurb } from 'src/components/About/Bio/types';
+
+export interface FetchBioRequest {
+    readonly type: typeof BIO_ACTIONS.FETCH_BIO_REQUEST;
+}
+
+export interface FetchBioError {
+    readonly type: typeof BIO_ACTIONS.FETCH_BIO_ERROR;
+}
+
+export interface FetchBioSuccess {
+    readonly type: typeof BIO_ACTIONS.FETCH_BIO_SUCCESS;
+    readonly bio: Blurb[];
+}
+
+type ActionTypes = FetchBioError | FetchBioRequest | FetchBioSuccess;
+
+export default ActionTypes;
