+++ web/src/components/About/Bio/index.ts
@@ -1,0 +1,7 @@
+import Bio from 'src/components/About/Bio/Bio';
+import * as bio from 'src/components/About/Bio/reducers';
+
+export const Component = Bio;
+export const reducers = {
+    bio: bio.bioReducer,
+};
