+++ web/src/components/About/Bio/reducers.ts
@@ -1,23 +1,23 @@
-import ABOUT_ACTIONS from 'src/components/About/actionTypeKeys';
-import ActionTypes from 'src/components/About/actionTypes';
-import { AboutStateShape } from 'src/components/About/types';
+import BIO_ACTIONS from 'src/components/About/Bio/actionTypeKeys';
+import ActionTypes from 'src/components/About/Bio/actionTypes';
+import { BioStateShape } from 'src/components/About/Bio/types';
 
-export const aboutReducer = (state: AboutStateShape = {
+export const bioReducer = (state: BioStateShape = {
     isFetching: false,
     bio: [],
 }, action: ActionTypes) => {
     switch (action.type) {
-        case ABOUT_ACTIONS.FETCH_BIO_REQUEST:
+        case BIO_ACTIONS.FETCH_BIO_REQUEST:
             return {
                 ...state,
                 isFetching: true,
             };
-        case ABOUT_ACTIONS.FETCH_BIO_ERROR:
+        case BIO_ACTIONS.FETCH_BIO_ERROR:
             return {
                 ...state,
                 isFetching: false,
             };
-        case ABOUT_ACTIONS.FETCH_BIO_SUCCESS:
+        case BIO_ACTIONS.FETCH_BIO_SUCCESS:
             return {
                 isFetching: false,
                 bio: action.bio,
