+++ web/src/components/About/Press/index.ts
@@ -1,0 +1,7 @@
+import Press from 'src/components/About/Press/Press';
+import { acclaimsListReducer } from 'src/components/About/Press/reducers';
+
+export const Component = Press;
+export const reducers = {
+    press_acclaimsList: acclaimsListReducer,
+};
