--- web/src/components/About/actionTypeKeys.ts
@@ -1,7 +1,0 @@
-enum ABOUT_ACTIONS {
-    FETCH_BIO_REQUEST = 'ABOUT--FETCH_BIO_REQUEST',
-    FETCH_BIO_SUCCESS = 'ABOUT--FETCH_BIO_SUCCESS',
-    FETCH_BIO_ERROR = 'ABOUT--FETCH_BIO_ERROR',
-}
-
-export default ABOUT_ACTIONS;
