+++ web/src/components/App/App.tsx
@@ -11,7 +11,6 @@ import { RequiredProps as AboutProps } from 'src/components/About/About';
 import { RequiredProps as ContactProps } from 'src/components/Contact/Contact';
 import { RequiredProps as HomeProps } from 'src/components/Home/Home';
 import { RequiredProps as MediaProps } from 'src/components/Media/Media';
-import { RequiredProps as PressProps } from 'src/components/Press/Press';
 import { RequiredProps as ScheduleProps } from 'src/components/Schedule/Schedule';
 
 import { Global } from '@emotion/core';
@@ -49,7 +48,6 @@ const About = () => register('about', import(/* webpackChunkName: 'about' */ 'sr
 const Contact = () => register('contact', import(/* webpackChunkName: 'contact' */ 'src/components/Contact'));
 const Home = () => register('home', import(/* webpackChunkName: 'home' */ 'src/components/Home'));
 const Media = () => register('media', import(/* webpackChunkName: 'media' */ 'src/components/Media'));
-const Press = () => register('press', import(/* webpackChunkName: 'press' */ 'src/components/Press'));
 const Schedule = () => register('schedule', import(/* webpackChunkName: 'schedule' */ 'src/components/Schedule'));
 const Page404 = () => register('page404', import(/* webpackChunkName: 'page404' */ 'src/components/Error'));
 
@@ -180,7 +178,7 @@ class App extends React.Component<AppProps, { homeBgLoaded: boolean; lastMatch?:
                                         <FadingContainer>
                                             <Switch location={this.props.location}>
                                                 <Route
-                                                    path="/about"
+                                                    path="/about/:about"
                                                     exact={true}
                                                     render={(childProps) => <AsyncComponent<AboutProps> moduleProvider={About} {...childProps} isMobile={matches} />}
                                                 />
@@ -193,11 +191,6 @@ class App extends React.Component<AppProps, { homeBgLoaded: boolean; lastMatch?:
                                                     path="/media/:media/:other*"
                                                     exact={true}
                                                     render={(childProps) => <AsyncComponent<MediaProps> moduleProvider={Media} {...childProps} isMobile={matches} />}
-                                                />
-                                                <Route
-                                                    path="/press"
-                                                    exact={true}
-                                                    render={(childProps) => <AsyncComponent<PressProps> moduleProvider={Press} {...childProps} isMobile={matches} />}
                                                 />
                                                 <Route
                                                     path="/schedule/:type?/:date?"
