--- web/src/components/Press/index.ts
@@ -1,7 +1,0 @@
-import Press from 'src/components/Press/Press';
-import { acclaimsListReducer } from 'src/components/Press/reducers';
-
-export const Component = Press;
-export const reducers = {
-    press_acclaimsList: acclaimsListReducer,
-};
