+++ web/src/types.ts
@@ -1,15 +1,15 @@
-import { AboutStateShape } from 'src/components/About/types';
+import { BioStateShape } from 'src/components/About/Bio/types';
+import { AcclaimsListStateShape } from 'src/components/About/Press/types';
 import { AudioPlaylistStateShape } from 'src/components/Media/Music/types';
 import { PhotoListReducerShape, PhotoViewerReducerShape } from 'src/components/Media/Photos/types';
 import { VideoPlayerStateShape, VideoPlaylistStateShape } from 'src/components/Media/Videos/types';
-import { AcclaimsListStateShape } from 'src/components/Press/types';
 import { ScheduleStateShape } from 'src/components/Schedule/types';
 
-import { aboutReducer } from 'src/components/About/reducers';
+import { bioReducer } from 'src/components/About/Bio/reducers';
+import { acclaimsListReducer } from 'src/components/About/Press/reducers';
 import { audioPlaylistReducer } from 'src/components/Media/Music/reducers';
 import { photoListReducer, photoViewerReducer } from 'src/components/Media/Photos/reducers';
 import { videoPlayerReducer, videoPlaylistReducer } from 'src/components/Media/Videos/reducers';
-import { acclaimsListReducer } from 'src/components/Press/reducers';
 import { scheduleReducer } from 'src/components/Schedule/reducers';
 
 import { Store } from 'redux';
@@ -17,7 +17,7 @@ import { Store } from 'redux';
 import { NavBarStateShape } from 'src/components/App/NavBar/types';
 
 export interface GlobalStateShape {
-    readonly about?: AboutStateShape;
+    readonly bio?: BioStateShape;
     readonly audio_playlist?: AudioPlaylistStateShape;
     readonly photo_list?: PhotoListReducerShape;
     readonly photo_viewer?: PhotoViewerReducerShape;
@@ -28,12 +28,12 @@ export interface GlobalStateShape {
     readonly navbar: NavBarStateShape;
 }
 
-export type AnyReducerType = typeof aboutReducer | typeof audioPlaylistReducer |
+export type AnyReducerType = typeof bioReducer | typeof audioPlaylistReducer |
     typeof photoListReducer | typeof photoViewerReducer | typeof acclaimsListReducer |
     typeof scheduleReducer | typeof videoPlayerReducer | typeof videoPlaylistReducer;
 
 export interface Reducers {
-    readonly about?: typeof aboutReducer;
+    readonly bio?: typeof bioReducer;
     readonly audio_playlist?: typeof audioPlaylistReducer;
     readonly photo_list?: typeof photoListReducer;
     readonly photo_viewer?: typeof photoViewerReducer;
